scrape-and-parse/AutoShake-feature-scrape-and-parse/python-resume/requirements.txt
python-dotenv
openai>=1.0.0
python-docx
